добавил проверку роли и владельца объявлений и комментов

+++ src/main/java/ru/skypro/homework/controllers/AdsController.java
@@ -2,12 +2,14 @@ package ru.skypro.homework.controllers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.AdDto;
 import ru.skypro.homework.dto.CreateOrUpdateAdDto;
 import ru.skypro.homework.dto.ExtendedAdDto;
 import ru.skypro.homework.model.Ad;
 import ru.skypro.homework.services.impl.AdsService;
+import ru.skypro.homework.services.impl.AuthServiceImpl;
 
 import java.util.List;
 
@@ -38,12 +40,19 @@ public class AdsController {
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @DeleteMapping("/ads/{id}")
     public void deleteAd(@PathVariable Integer id) {
-        adsService.deleteAd(id);
+        if (adsService.getByAdId(id).getUser().equals(AuthServiceImpl.getAuthUser())
+                || adsService.getByAdId(id).getUser().getRole().name().equals("ADMIN")) {
+            adsService.deleteAd(id);
+        }
     }
 
     @PatchMapping("/ads/{id}")
     public AdDto updateAd(@RequestBody CreateOrUpdateAdDto dto, @PathVariable Integer id) {
-        return adsService.updateAd(dto, id);
+        if (adsService.getByAdId(id).getUser().equals(AuthServiceImpl.getAuthUser())
+                || adsService.getByAdId(id).getUser().getRole().name().equals("ADMIN")) {
+            return adsService.updateAd(dto, id);
+        }
+        throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/ads/me")
