+++ src/main/java/ru/skypro/homework/controllers/CommentController.java
@@ -1,7 +1,10 @@
 package ru.skypro.homework.controllers;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.HttpClientErrorException;
 import ru.skypro.homework.dto.CommentDto;
+import ru.skypro.homework.services.impl.AuthServiceImpl;
 import ru.skypro.homework.services.impl.CommentService;
 
 import java.util.List;
@@ -28,11 +31,18 @@ public class CommentController {
 
     @DeleteMapping("/{adId}/comments/{commentId}")
     public void deleteComment(@PathVariable Integer adId, @PathVariable Integer commentId) {
-        commentService.deleteComment(adId, commentId);
+        if (commentService.getByCommentId(commentId).getUser().equals(AuthServiceImpl.getAuthUser())
+                || commentService.getByCommentId(commentId).getUser().getRole().name().equals("ADMIN")){
+            commentService.deleteComment(adId, commentId);
+        }
     }
 
     @PatchMapping("/{adId}/comments/{commentId}")
     public CommentDto updateComment(@RequestBody String text, @PathVariable Integer adId, @PathVariable Integer commentId) {
-        return commentService.updateComment(text, adId, commentId);
+        if (commentService.getByCommentId(commentId).getUser().equals(AuthServiceImpl.getAuthUser())
+                || commentService.getByCommentId(commentId).getUser().getRole().name().equals("ADMIN")){
+            return commentService.updateComment(text, adId, commentId);
+        }
+        throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
     }
 }
